requirements.txt
google-generativeai==0.3.1
streamlit==1.29.0
python-dotenv==1.1.1